Specified return object.
